refactor: move non-number specific keyboard functions to Input instead of InputNumber
