+++ packages/ui-app/src/InputNumber.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import isString from '@polkadot/util/is/string';
 
 import classes from './util/classes';
-import Input from './Input';
+import Input, { KEYS, KEYS_PRE, isCopy, isCut, isDuplicateDecimalPoint, isPaste, isSelectAll } from './Input';
 import translate from './translate';
 
 type Props = BareProps & I18nProps & {
@@ -31,50 +31,7 @@ type State = {
 // chain specification bit length
 const BIT_LENGTH_128 = 128;
 
-// note: KeyboardEvent.keyCode and KeyboardEvent.which are deprecated
-const KEYS = {
-  A: 'a',
-  ALT: 'Alt',
-  ARROW_LEFT: 'ArrowLeft',
-  ARROW_RIGHT: 'ArrowRight',
-  BACKSPACE: 'Backspace',
-  C: 'c',
-  CMD: 'Meta',
-  CTRL: 'Control',
-  DECIMAL_POINT: '.',
-  ENTER: 'Enter',
-  ESCAPE: 'Escape',
-  TAB: 'Tab',
-  V: 'v',
-  X: 'x'
-};
-
 const KEYS_ALLOWED: Array<any> = [KEYS.ARROW_LEFT, KEYS.ARROW_RIGHT, KEYS.BACKSPACE, KEYS.DECIMAL_POINT, KEYS.ENTER, KEYS.ESCAPE, KEYS.TAB];
-
-const KEYS_PRE: Array<any> = [KEYS.ALT, KEYS.CMD, KEYS.CTRL];
-
-// reference: degrade key to keyCode for cross-browser compatibility https://www.w3schools.com/jsref/event_key_keycode.asp
-const isCopy = (key: string, isPreKeyDown: boolean): boolean =>
-  isPreKeyDown && key === KEYS.C;
-
-const isCut = (key: string, isPreKeyDown: boolean): boolean =>
-  isPreKeyDown && key === KEYS.X;
-
-const regexDecimalPoint = /[\.]/gi;
-
-const isDuplicateDecimalPoint = (key: string, value: string): boolean => {
-  const inputValue: string = value;
-  const didPressDecimalPoint: boolean = key === KEYS.DECIMAL_POINT;
-  const foundExistingDecimalPoint: boolean = inputValue.match(regexDecimalPoint) ? true : false;
-
-  return didPressDecimalPoint && foundExistingDecimalPoint;
-};
-
-const isPaste = (key: string, isPreKeyDown: boolean): boolean =>
-  isPreKeyDown && key === KEYS.V;
-
-const isSelectAll = (key: string, isPreKeyDown: boolean): boolean =>
-  isPreKeyDown && key === KEYS.A;
 
 class InputNumber extends React.PureComponent<Props, State> {
   state: State = {
